requirements.txt
# Core
flask==3.0.0
gunicorn==21.2.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# Google Cloud
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0
google-auth==2.27.0

# Environment
python-dotenv==1.0.0
